requirements.txt
Django==1.11
djangorestframework==3.8.2
pytz==2018.4
bs4
requests==2.18.4

